shop/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.2
django-bootstrap5==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
